refactor(settings): extract proxy registry helpers
